requirements.txt
py-cord
jishaku
discord-pretty-help
motor